♻️ refactor :  🏗 folder :  fixed된 컴포넌트들 공간 잘 차지시켜놓기 -> 헤더 및 바텀 관련 컴포넌트들 전부 정리 완
