+++ src/app/(Main)/feed/page.tsx
@@ -3,13 +3,10 @@
 import { QueryAsyncBoundary } from '@suspensive/react-query';
 
 import FilterSheet from './components/FilterSheet';
-import FilterTab from './components/FilterTab';
 import FeedContent from './FeedContent';
 import { RejectedFallback } from '@/components/ErrorBoundary';
-import { MainHeader } from '@/components/Header';
 import { Loading } from '@/components/Loading';
 import { PageAnimation } from '@/components/PageAnimation';
-import { Spacing } from '@/components/Spacing';
 
 export default function Feed() {
   return (
@@ -18,10 +15,6 @@ export default function Feed() {
       pendingFallback={<Loading className="h-[calc(100dvh-178px)]" />}
     >
       <PageAnimation>
-        <MainHeader>
-          <FilterTab />
-        </MainHeader>
-        <Spacing size={50} />
         <FeedContent />
         <FilterSheet />
       </PageAnimation>
