+++ src/app/(Main)/talk/page.tsx
@@ -1,16 +1,12 @@
 import TalkSection from './components/TalkSection';
 import TitleSection from './components/TitleSection';
 import { PageAnimation } from '@/components/PageAnimation';
-import { Spacing } from '@/components/Spacing';
 
 export default function Talk() {
   return (
-    <>
-      <PageAnimation className="flex flex-1 flex-col items-center justify-center">
-        <TitleSection />
-        <TalkSection />
-      </PageAnimation>
-      <Spacing size={100} />
-    </>
+    <PageAnimation className="flex flex-1 flex-col items-center justify-center">
+      <TitleSection />
+      <TalkSection />
+    </PageAnimation>
   );
 }
