+++ src/app/(Sub)/layout.tsx
@@ -1,11 +1,5 @@
-import { Spacing } from '@/components/Spacing';
 import { StrictPropsWithChildren } from '@/types';
 
 export default function SubLayout({ children }: StrictPropsWithChildren) {
-  return (
-    <>
-      <Spacing size={60} />
-      {children}
-    </>
-  );
+  return <>{children}</>;
 }
