+++ src/components/Header/index.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren } from 'react';
 
 import { CloseButton, MenuButton } from './HeaderButton';
-import Tab from './Tab';
+import { Spacing } from '../Spacing';
 
 interface Header {
   title?: string;
@@ -16,20 +16,16 @@ export default function Header({
   children,
 }: PropsWithChildren<Header>) {
   return (
-    <div className="fixed inset-x-0 top-0 z-10 mx-auto max-w-440 bg-white">
-      <div className="flex h-48 items-center justify-between gap-12 px-4 pt-8">
-        {close ? <CloseButton /> : <div className="w-4" />}
-        <h4 className="flex flex-1">{title}</h4>
-        <div className="flex">{menu && <MenuButton />}</div>
+    <>
+      <Spacing size={48} />
+      <div className="fixed inset-x-0 top-0 z-10 mx-auto max-w-440 bg-white">
+        <div className="flex h-48 items-center justify-between gap-12 px-4 pt-8">
+          {close ? <CloseButton /> : <div className="w-4" />}
+          <h4 className="flex flex-1">{title}</h4>
+          <div className="flex">{menu && <MenuButton />}</div>
+        </div>
+        {children}
       </div>
-      {children}
-    </div>
+    </>
   );
 }
-
-export const MainHeader = ({ children }: PropsWithChildren) => (
-  <Header title="PosePicker" menu={true}>
-    <Tab />
-    {children}
-  </Header>
-);
